put liveData and stateFlow methods in their own files (also implements getStateFlow with async key)
